feat(model): The player model now has game-accurate proportions
In game every 2nd layer skin part has an inflate value of 0.25 with the exception of the hat model which has an inflate value of 0.5. The inflate value adds the <inflate> amount of width on all sides which means if a model is 3 pixels wide and has an inflate value of 0.5 it will be 4 pixels wide. since in these models here 1 is equal to 8 pixels in game that makes it so the head is 9/8 (1.125) pixels wide and every other model is 1/16 (0.0625) pixels wider than the base model.